move ripple utilities to utils.js
